requirements.txt
APScheduler==3.9.1
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
dataclasses-json==0.5.7
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
isodate==0.6.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-youtube==0.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing-inspect==0.7.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
